c32fac079f8b887b9a6e274a2eb3b47c5ee01d93 refactor: more `CGovernanceVote` cleanups (UdjinM6)
efc8c99139659695ebfff2139645f92b8055b2f4 refactor: Introduce `get_chain_conf_names` (UdjinM6)
519966efff46e8582dee69c1e233ca0c29b67bf9 refactor: Refactor `CCbTx` python class to match c++ declaration (UdjinM6)
38a0b5ac04cfef2421457dee8dc44b887ef73d41 refactor: Move `HasChainLock()` calls out of loops (UdjinM6)
6370c7a9e5bfadbe7d55bef3b50df61103bbaa39 refactor: Tidy up `CGovernanceVote` a bit (UdjinM6)
ae675d5314b846ea0a1f0d807f9ffb9aeca5527f refactor: Drop redundant `SplitBy` (UdjinM6)
88c3a13674c720db348c7d73e4b771a96f9cab51 refactor: Use `switch..case` instead of `if` in `AvailableCoins` (UdjinM6)
38b304c7b0f2e91a82277bfcc9b879ab75cdbf20 refactor: Make bls worker aggregator batch size const (rename it accordingly) (UdjinM6)
0cf7d4a65efb05c6465abd1298134bc60d4a1420 refactor: Turn `ReadResult` into enum class (UdjinM6)
d9aab0fa6045572036a740fcdb7cfb09b0166488 refactor: Rename `masternode_state_t` and turn it into enum class (UdjinM6)

Pull request description:

  ## Issue being fixed or feature implemented

  ## What was done?

  ## How Has This Been Tested?

  ## Breaking Changes

  ## Checklist:
  - [ ] I have performed a self-review of my own code
  - [ ] I have commented my code, particularly in hard-to-understand areas
  - [ ] I have added or updated relevant unit/integration/functional/e2e tests
  - [ ] I have made corresponding changes to the documentation
  - [ ] I have assigned this pull request to a milestone _(for repository code-owners and collaborators only)_

ACKs for top commit:
  knst:
    utACK c32fac079f8b887b9a6e274a2eb3b47c5ee01d93
  PastaPastaPasta:
    utACK c32fac079f8b887b9a6e274a2eb3b47c5ee01d93

Tree-SHA512: 7a33d2d535f73c858cd01a442fb28793b1ea185a2eb50d6bce86b44be1012b7427649e3496647b4a0a7d7816dbbb50f138fa701d203c2aa1e7a515c9c9a0e931

+++ src/interfaces/node.h
@@ -39,7 +39,7 @@ enum class TransactionError;
 struct CNodeStateStats;
 struct NodeContext;
 
-enum vote_signal_enum_t : uint8_t;
+enum vote_signal_enum_t : int;
 
 namespace interfaces {
 class Handler;
